done exercise with gorilla/mux

+++ rest-api/api/models/comments.go
@@ -4,11 +4,11 @@ import "time"
 
 //Comment is an...
 type Comment struct {
-	ID      uint32 `gorm:"primary_key;auto_increment" json:"id"`
+	ID      uint64 `gorm:"primary_key;auto_increment" json:"id"`
 	Comment string `gorm:"size:255 ;not null" json:"comments"`
-	UserID  uint32 `gorm:"not null" json:"user_id"`
+	UserID  uint64 `gorm:"not null" json:"user_id"`
 
-	PostID uint32 `gorm:"not null" json:"post_id"`
+	PostID uint64 `gorm:"not null" json:"post_id"`
 
 	CreatedAt time.Time `gorm:"type:timestamp(0);default:current_timestamp();not null" json:"created_at"`
 	UpdatedAt time.Time `gorm:"type:timestamp(0);default:current_timestamp();not null" json:"updated_at"`
